roles/pythons/files/requirements.txt
python-telegram-bot[job-queue]==21.1.1
pydantic==2.7.1
isort==5.13.2
black==24.4.2